Wire up new message notifications
Notify if a message is new to us, unread, at most seven days old, and in the inbox folder.
In the future, it would be nice to make this smarter and/or configurable. For now, it's important to be conservative: we definitely don't want to send notifications when a message enters sent, deleted, junk, or similar folders.